python/requirements.txt
Django==5.0.2
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
Pillow==11.1.0
django-filter==23.5
django-widget-tweaks==1.5.0
python-dotenv==1.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1 